keyboard shortcuts

+++ src/main/java/com/niklim/clicktrace/view/editor/action/screenshot/DeleteScreenShotActionListener.java
@@ -6,6 +6,7 @@ import java.awt.event.ActionListener;
 import javax.swing.JOptionPane;
 
 import com.google.inject.Inject;
+import com.niklim.clicktrace.controller.ActiveSession;
 import com.niklim.clicktrace.controller.Controller;
 import com.niklim.clicktrace.view.editor.Editor;
 
@@ -14,10 +15,17 @@ public class DeleteScreenShotActionListener implements ActionListener {
 	private Editor editor;
 	@Inject
 	private Controller controller;
+	@Inject
+	private ActiveSession activeSesssion;
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		int answer = JOptionPane.showConfirmDialog(editor.getFrame(), "Are you sure?", "", JOptionPane.OK_CANCEL_OPTION);
+		if (activeSesssion.getSession() == null || activeSesssion.getActiveShot() == null) {
+			return;
+		}
+		String screenShot = activeSesssion.getActiveShot().toString();
+		int answer = JOptionPane.showConfirmDialog(editor.getFrame(), "Are you sure to delete '"
+				+ screenShot + "' screenshot?", "", JOptionPane.OK_CANCEL_OPTION);
 		if (answer == JOptionPane.OK_OPTION) {
 			controller.deleteActiveScreenShot();
 		}
